requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
bleach==5.0.1
blis==0.7.9
captum==0.6.0
catalogue==2.0.8
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
confection==0.0.3
confusable-homoglyphs==3.2.0
contourpy==1.0.6
cycler==0.11.0
cymem==2.0.7
decorator==5.1.1
Django==4.0.8
django-classy-tags==4.0.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.1
django-guardian==2.4.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-nyt==1.2.4
django-registration==3.3
django-sekizai==4.0.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-nested-routers==0.93.4
enum-compat==0.0.3
filelock==3.9.0
fonttools==4.38.0
fst-pso==1.8.1
future==0.18.2
FuzzyTM==2.0.5
gensim==4.3.0
huggingface-hub==0.11.1
idna==3.4
isodate==0.6.1
Jinja2==3.1.2
jsonpath-python==1.0.6
kiwisolver==1.4.4
kombu==5.2.4
langcodes==3.3.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.6.2
miniful==0.0.6
murmurhash==1.0.9
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==22.0
Pairtree==0.8.1
pandas==1.5.2
pathy==0.10.1
Pillow==9.3.0
ply==3.11
preshed==3.0.8
prompt-toolkit==3.0.36
psutil==5.9.4
pydantic==1.10.3
pyFUME==0.2.25
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
PyYAML==6.0
rdflib==6.2.0
regex==2022.10.31
requests==2.28.1
scipy==1.9.3
simpful==2.9.0
six==1.16.0
smart-open==6.3.0
sorl-thumbnail==12.9.0
spacy==3.4.4
spacy-legacy==3.0.11
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.5
thinc==8.1.6
tinycss2==1.1.1
tokenizers==0.13.2
torch==1.13.1
torch-model-archiver==0.7.0
torchvision==0.14.1
tqdm==4.64.1
transformers==4.25.1
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.13
vine==5.0.0
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
wiki==0.9
Wikidata==0.7.0
